third_party/WebKit/LayoutTests/inspector/elements/shadow/shadow-root.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<script>

function test()
{
    ElementsTestRunner.expandElementsTree(function() {
        var container = ElementsTestRunner.expandedNodeWithId("container");
        ElementsTestRunner.dumpElementsTree(container);
        TestRunner.completeTest();
    });
}
</script>
</head>

<body onload="runTest()">

<p id="description">This test verifies that author shadow root's #document-fragment is displayed and user-agent one is hidden by default.</p>

<div id="container">
    <div id="test1"></div>
    <div id="test2">only test</div>
    <div id="test3">with <span>elements</span></div>
    <input type="text" value="Test">
</div>

<script>
test1.createShadowRoot();
test2.attachShadow({ mode: "open" });
test3.attachShadow({ mode: "closed" });
</script>

</body>
</html>
